chore: remove dead legacySideEffects hydration modules
- Delete unused runMeshcoreMountHydration and runMeshtasticDbHydration
  (superseded by hydrateIdentityStoresFromDb and runtime reload paths).
- Extract getMeshtasticMessageLoadLimit to meshtasticMessageLoadLimit.ts.
- Move mirrorMqttStatusToConnection into connectionStore; drop
  legacySideEffects/ directory.
- Update AGENTS.md to document canonical identity-scoped hydration ([#375]).

Wire subscription modules (meshtasticLegacyWireSubscriptions,
meshcoreLegacyConnEvents) remain until protocol event migration.
